updated product status dto
Datatype changed from int to long
switched to object type (Long) from primitive type (long)

+++ src/main/java/com/cart/dto/ProductStatusDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class ProductStatusDTO {
-    private int ordered;
-    private int discarded;
-    private int active;
+    private Long ordered;
+    private Long discarded;
+    private Long active;
 }
